Rename `serverSidePropsMonitorComponent.tsx` to `connectionMonitorComponent.tsx`.
