+++ lib/connectionMonitorComponent.tsx
